% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkExtension.R
\name{checkExtension}
\alias{checkExtension}
\title{Checks if file extension matches}
\usage{
checkExtension(file, exp_ext)
}
\arguments{
\item{file}{The file of which the extension needs to be checked.}

\item{exp_ext}{The expected extension of the file to be checked. This
extension file can be both with or without leading "." For example,
inputing .csv or csv will give similar results.}
}
\value{
A TRUE value if the extension matches. And a FALSE value if
the extension does not match.
}
\description{
Checks if file extension matches
}
\examples{
checkExtension(file = "testdata/test1.csv", exp_ext = "csv")
}
